S4 class for performing a landmarking analysis

S4 class for performing a landmarking analysis

Slots

landmarks

A numeric vector of landmark times.

data_static

A data frame containing subject ids, static covariates,

data_dynamic

Data frame in long format with subject ids, measurement values, measurement times and measurement name.

event_indicator

Name of the column indicating event or censoring.

dynamic_covariates

Names of dynamic (time-varying) covariates. Character vector.

ids

Name of the column indicating subject ids.

times

Name of the column indicating observation time in data_dynamic.

measurements

Name of the column indicating measurement values in data_dynamic.

dynamic_covariate_names

Name of the column indicating names of the dynamic covariates in data_dynamic.

event_time

Name of the column indicating time of the event/censoring.

risk_sets

List of indices.

longitudinal_fits

List of model fits for the specified landmark times and biomarkers.

longitudinal_predictions

List of model predictions for the specified landmark times and biomarkers.

survival_fits

List of survival model fits at each of the specified landmark times.